+++ HTTPCore/trunk/HTTPCore/encoders.h
@@ -8,18 +8,18 @@ public:
 	encoders();
 	~encoders();
 	char* decodebase64(char *output, const char *input);
-	char* encodebase64(char *output, const char *input, unsigned int len);	
+	char* encodebase64(char *output, const char *input, size_t len);	
 	char *CreateDigestAuth(char *AuthenticationHeader, const char *lpUsername, const char *lpPassword, const char *method,const char *uri, int counter);
 	char *GetNTLMBase64Packet1(char*destination);
 	char *GetNTLMBase64Packet3(char*destination, const char* NTLMresponse, const char *lpUsername, const char* lpPassword);
-	unsigned char* GetMD2BinaryHash(char *output, const char *data, unsigned int len);
-	char* GetMD2TextHash(char *output, const char *data, unsigned int len);
-	unsigned char* GetMD4BinaryHash(char *output, const char *data, unsigned int len);
-	char* GetMD4TextHash(char *output, const char *data, unsigned int len);
-	unsigned char* GetMD5BinaryHash(char *output, const char *data, unsigned int len);
-	char* GetMD5TextHash(char *output, const char *data, unsigned int len);
-	unsigned char* GetSHA1BinaryHash(char *output, const char *data, unsigned int len);
-	char* GetSHA1TextHash(char *output, const char *data, unsigned int len);
+	unsigned char* GetMD2BinaryHash(char *output, const char *data, size_t len);
+	char* GetMD2TextHash(char *output, const char *data, size_t len);
+	unsigned char* GetMD4BinaryHash(char *output, const char *data, size_t len);
+	char* GetMD4TextHash(char *output, const char *data, size_t len);
+	unsigned char* GetMD5BinaryHash(char *output, const char *data, size_t len);
+	char* GetMD5TextHash(char *output, const char *data, size_t len);
+	unsigned char* GetSHA1BinaryHash(char *output, const char *data, size_t len);
+	char* GetSHA1TextHash(char *output, const char *data, size_t len);
 
 };
